Fix lnd safe shutdown.
lnd needs to be able to shutdown fully, and currently because it's a child process
of the main node process, it is killed when the app is killed and isn't given a
chance to fully tear down. Now, we start lnd as a detached process that can die
at its own pace when the app is closed.

A potential downside is that the lnd logs no longer appear alongside the app logs.
However, the lnd logs are still accessible in the app's data directory.